% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pcctc_project.R
\name{create_pcctc_project}
\alias{create_pcctc_project}
\alias{add_project_directory}
\title{Create PCCTC Project}
\usage{
create_pcctc_project(
  path,
  path_data,
  renv = TRUE,
  overwrite = NA,
  open = interactive()
)

add_project_directory(dir_name, overwrite = NA)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created.}

\item{path_data}{A path. The directory where the secure data exist. Default is
\code{NULL}.  When supplied, a symbolic link to data folder will be created.}

\item{renv}{Logical indicating whether to add renv to a project.
Default is \code{TRUE}. When \code{NA} user is asked interactively for preference.}

\item{overwrite}{Logical indicating whether to overwrite existing files
if they exist.  Options are
\code{TRUE}, \code{FALSE}, and \code{NA} (aka ask interactively).  Default is \code{NA}}

\item{open}{Logical indicating whether to open new project in fresh RStudio
session}

\item{dir_name}{name of directory to add to project folder.}
}
\description{
A typical project directory will look like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{├── .git
├── README.md
├── _metadata
│   ├── _env.yaml
│   ├── european-urology.csl
│   └── references.bib
├── 01-data-setup
│   ├── README.md
│   ├── setup0_data-setup_medidata.qmd
│   ├── setup1_data-setup.qmd
│   ├── setup2_data-setup_progression.qmd
│   └── derived-variables_data-setup.xlsx
├── 02-metrics
│   ├── README.md
│   ├── analysis1_metrics.qmd
│   └── report1_metrics.qmd
├── 03-data-review
│   ├── README.md
│   └── analysis1_data-reviews-pointblank.qmd
├── 04-visit-reports
│   ├── README.md
│   └── report1_visit-reports.qmd
├── 05-trial-aims
│   ├── README.md
│   ├── analysis1_trial-aims-consort.qmd
│   ├── analysis2_trial-aims-safety.qmd
│   ├── analysis3_trial-aims-primary-aims.qmd
│   ├── report1_trial-aims-consort.qmd
│   ├── report2_trial-aims-safety.qmd
│   └── report3_trial-aims-primary-aims.qmd
}\if{html}{\out{</div>}}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# add example
\dontshow{\}) # examplesIf}
}
\author{
Daniel D. Sjoberg
}
