% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croquet-package.R
\docType{package}
\name{croquet-package}
\alias{croquet}
\alias{croquet-package}
\title{croquet: Miscellaneous Operations For PCCTC}
\description{
Miscellaneous operations for PCCTC, a Clinical Research Organization (CRO).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pcctc/croquet}
  \item \url{https://pcctc.github.io/croquet/}
  \item Report bugs at \url{https://github.com/pcctc/croquet/issues}
}

}
\author{
\strong{Maintainer}: Shannon Pileggi \email{pileggis@mskcc.org}

Authors:
\itemize{
  \item Travis Gerke \email{gerket@mskcc.org}
}

}
\keyword{internal}
