% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find duplicate observations in a data frame by a identifier.}
\usage{
find_duplicates(data, id)
}
\arguments{
\item{data}{a data frame}

\item{id}{a variable containing the supposedly unique identifier}
}
\value{
a tibble with non-uniquely identified rows
}
\description{
Find duplicate observations in a data frame by a identifier.
}
\examples{
x <- data.frame(subject = c("a", "a", "b", "c", "c"))
y <- data.frame(subject = LETTERS[1:5])
find_duplicates(x, subject)
find_duplicates(y, subject)
}
