% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_split_named.R
\name{group_split_named}
\alias{group_split_named}
\title{Splits data frame by groups into a named list}
\usage{
group_split_named(.tbl, ...)
}
\arguments{
\item{.tbl}{a data frame}

\item{...}{Grouping specification, forwarded to group_by()}
}
\value{
a named list
}
\description{
Behaves similarly to dplyr::group_split(), with the addition of
named list elements. Borrowed from Romain Francois in this dplyr issue:
https://github.com/tidyverse/dplyr/issues/4223
}
\examples{
\dontrun{
mtcars \%>\% group_split_named(am)
mtcars \%>\% group_split_named(cyl, am)
}


}
