% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_data.R
\name{here_data}
\alias{here_data}
\alias{get_data_date}
\title{Path to Data Directory}
\usage{
here_data(...)

get_data_date()
}
\arguments{
\item{...}{Path components below the project root, can be empty.
Each argument should be a string containing one or more path components.}
}
\value{
a path
}
\description{
Path to Data Directory
}
