% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_sheet.R
\name{labelled_sheet}
\alias{labelled_sheet}
\title{A hidden function to add labelled data sheet to an excel workbook through openxlsx}
\usage{
labelled_sheet(data, sheet_name = NULL, wrkbk = wb, start_row = 1L)
}
\arguments{
\item{data}{labelled data to add to sheet}

\item{sheet_name}{optional sheet name; if none provided, sheet will be assigned name
of input data set}

\item{wrkbk}{workbook name, defaults to wb}

\item{start_row}{integer row position where the labels must be placed, \code{data} is placed at start_row+1, defaults to 1L}
}
\value{
a list containing worksheet definition; the function modifies the object passed to the \code{wb} argument.
}
\description{
You can add additional styling options to sheet with the openxlsx::add functions
after executed.
}
\examples{
\dontrun{
options("openxlsx.dateFormat" = "yyyy-mm-dd")
dat_labelled <- tibble::tibble(
  var_1 = 1:3,
  var_2 = LETTERS[1:3],
  var_3 = Sys.Date() - 0:2
  ) \%>\%
  labelled::set_variable_labels(
    var_1 = "Variable 1 (numbers)",
    var_2 = "Variable 2 (letters)",
    var_3 = "Variable 3 (date)"
  )
wb <- createWorkbook()
wb <- labelled_sheet(dat_labelled, sheet_name = "example sheet", wrkbk = wb)
saveWorkbook(wb, "checkwb.xlsx")
}
}
