% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_raw_medidata.R
\name{prepare_raw_medidata}
\alias{prepare_raw_medidata}
\title{Prepare Raw Medidata}
\usage{
prepare_raw_medidata(x, rename_fn = toupper, retain_datetime = NULL)
}
\arguments{
\item{x}{list of Medidata data frames, often imported with \code{read_sas_batch()}}

\item{rename_fn}{function passed to \code{dpylr::rename_with(.fn=)}. Default is \code{toupper}.}

\item{retain_datetime}{character vector of column names that will not be converted to dates.}
}
\value{
list of data frames
}
\description{
Function performs initial cleaning of a list of Medidata data frames.
\itemize{
\item Convert empty strings to \code{NA}.
\item Convert all date-time columns to dates.
\item Any dates imputed to the year 1900 are made \code{NA}.
\item All variables are made upper/lower case (depending on the function passed in \verb{rename_fn=}).
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_sas_batch(x) |>
  prepare_raw_medidata()
\dontshow{\}) # examplesIf}
}
