% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_templates}
\alias{project_templates}
\title{PCCTC project templates}
\format{
A named list containing the PCCTC project templates.
}
\usage{
project_templates
}
\description{
The \code{project_templates} object defines the contents of the PCCTC project
templates used in \code{create_pcctc_project()} and \code{use_pcctc_file()}.
}
\keyword{datasets}
