% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_labels.R
\name{retrieve_labels}
\alias{retrieve_labels}
\title{Retrieve labels from labelled data}
\usage{
retrieve_labels(data, wide = TRUE)
}
\arguments{
\item{data}{labelled data frame}

\item{wide}{format of return: wide (default) returns a tibble with a single row
that contains the variable labels; otherwise, the full data dictionary is returned}
}
\value{
a tibble
}
\description{
Retrieve labels from labelled data
}
\examples{

dat_labelled <- tibble::tibble(
  var_1 = 1:3,
  var_2 = LETTERS[1:3],
  var_3 = Sys.Date() - 0:2
  ) \%>\%
  labelled::set_variable_labels(
    var_1 = "Variable 1 (numbers)",
    var_2 = "Variable 2 (letters)",
    var_3 = "Variable 3 (date)"
  )

retrieve_labels(dat_labelled)
retrieve_labels(dat_labelled, wide = FALSE)

}
