% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_derived_variable_labels.R
\name{set_derived_variable_labels}
\alias{set_derived_variable_labels}
\title{Apply variable labels to data frame}
\usage{
set_derived_variable_labels(data, df_name, path, drop = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{df_name}{string indicating the name of the data frame to apply labels to.
If not specified, we'll do our best to determine the name of the passed data frame.}

\item{path}{Path to CSV file}

\item{drop}{Logical indicating whether to drop unlabeled variables}
}
\description{
Takes labels from the Derived Variables CSV file and applies them to the
passed data frame.
The excel sheet must have columns \code{"df_name"}, \code{"df_label"}, \code{"var_name"}, and \code{"var_label"}.
The returned data frame is also labelled with the value in \code{"df_label"}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
trial \%>\%
  set_derived_variable_labels("derived_variables_sjoberg.csv")
\dontshow{\}) # examplesIf}
}
