% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pcctc_file.R
\name{use_pcctc_file}
\alias{use_pcctc_file}
\alias{use_pcctc_gitignore}
\alias{use_pcctc_setup}
\alias{use_pcctc_analysis}
\alias{use_pcctc_report}
\title{Write PCCTC Project File}
\usage{
use_pcctc_file(name = NULL, filename = NULL, open = interactive())

use_pcctc_gitignore(filename = NULL)

use_pcctc_setup(filename = NULL)

use_pcctc_analysis(filename = NULL)

use_pcctc_report(filename = NULL)
}
\arguments{
\item{name}{Name of file to write.  Not sure of the files available to you?
Run the function without specifying a name, and all files available within the
template will print.}

\item{filename}{Optional argument to specify the name of the file to be written.}

\item{open}{Logical indicating whether to open new project in fresh RStudio
session}
}
\description{
You may use \code{use_pcctc_file()} to write a single file from the PCCTC project template.
The functions \code{use_pcctc_gitignore()} and friends are shortcuts for
\code{use_pcctc_file("gitignore")}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# create gitignore file
use_pcctc_file("gitignore")
use_pcctc_gitignore()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=create_pcctc_project]{create_pcctc_project()}}
}
