% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pcctc_rstudio_prefs.R
\name{use_pcctc_rstudio_prefs}
\alias{use_pcctc_rstudio_prefs}
\title{Use Data Science RStudio Preferences}
\usage{
use_pcctc_rstudio_prefs(profile = tolower(Sys.info()[["user"]]))
}
\arguments{
\item{profile}{Used to set additional preferences saved under specified profile.
Default is your current system username. If no profile exists, it is ignored.}
}
\description{
The function wraps \code{rstudio.prefs::use_rstudio_prefs()} and sets the following
preferences in RStudio.\tabular{ll}{
   \strong{Preference} \tab \strong{Value} \cr
   always_save_history \tab FALSE \cr
   graphics_backend \tab ragg \cr
   load_workspace \tab FALSE \cr
   show_hidden_files \tab TRUE \cr
   show_line_numbers \tab TRUE \cr
   margin_column \tab 80 \cr
   save_workspace \tab never \cr
   rainbow_parentheses \tab TRUE \cr
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
use_pcctc_rstudio_prefs()
\dontshow{\}) # examplesIf}
}
