% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pcctc_project.R
\name{create_pcctc_project}
\alias{create_pcctc_project}
\alias{add_project_subdirectory}
\title{Create PCCTC Project}
\usage{
create_pcctc_project(path, renv = TRUE, overwrite = NA)

add_project_subdirectory(path, renv = TRUE, open = TRUE, overwrite = NA)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created.}

\item{renv}{Logical indicating whether to add renv to a project.
Default is \code{TRUE}. When \code{NA} user is asked interactively for preference.}

\item{overwrite}{Logical indicating whether to overwrite existing files
if they exist.  Options are
\code{TRUE}, \code{FALSE}, and \code{NA} (aka ask interactively).  Default is \code{NA}}

\item{open}{Logical indicating whether to open new project in fresh RStudio
session}
}
\description{
Create PCCTC Project
}
\examples{
# add example
}
