% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pcctc_project.R
\name{create_pcctc_project}
\alias{create_pcctc_project}
\alias{add_project_directory}
\title{Create PCCTC Project}
\usage{
create_pcctc_project(
  path,
  path_data,
  renv = TRUE,
  overwrite = NA,
  open = interactive()
)

add_project_directory(dir_name, overwrite = NA)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created.}

\item{path_data}{A path. The directory where the secure data exist. Default is
\code{NULL}.  When supplied, a symbolic link to data folder will be created.}

\item{renv}{Logical indicating whether to add renv to a project.
Default is \code{TRUE}. When \code{NA} user is asked interactively for preference.}

\item{overwrite}{Logical indicating whether to overwrite existing files
if they exist.  Options are
\code{TRUE}, \code{FALSE}, and \code{NA} (aka ask interactively).  Default is \code{NA}}

\item{open}{Logical indicating whether to open new project in fresh RStudio
session}

\item{dir_name}{name of directory to add to project folder.}
}
\description{
Create PCCTC Project
}
\examples{
# add example
}
\author{
Daniel D. Sjoberg
}
