% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find duplicate observations in a data frame by a identifier.}
\usage{
find_duplicates(x, ..., sort = FALSE, name = "n")
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to group
by.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}
}
\description{
Find duplicate observations in a data frame by a identifier.
}
\examples{
x <- data.frame(subject = c("a", "a", "b", "c", "c"))
find_duplicates(x, subject)
find_duplicates(mtcars, vs, am)

}
