% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_helpers}
\alias{select_helpers}
\alias{all_uppercase}
\alias{all_lowercase}
\title{Select Helpers}
\usage{
all_uppercase()

all_lowercase()
}
\description{
\itemize{
\item \code{all_uppercase()} select all columns with names that are upper case
\item \code{all_lowercase()} select all columns with names that are lower case
}
}
\examples{
df <- data.frame(ONE = 1, two = 2)

df |> select(all_uppercase())
df |> select(all_lowercase())
}
