% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtsummary_pcctc.R
\name{theme_gtsummary_pcctc}
\alias{theme_gtsummary_pcctc}
\title{Set PCCTC gtsummary theme}
\usage{
theme_gtsummary_pcctc(font_size = NULL)
}
\arguments{
\item{font_size}{Numeric font size for compact theme.
Default is 13 for gt tables, and 8 for all other output types}
}
\description{
Visit the \href{https://www.danieldsjoberg.com/gtsummary/articles/themes.html#writing-themes-1}{gtsummary themes vignette}
for a full list of preferences that can be set.
}
\examples{
theme_gtsummary_pcctc()
}
\concept{gtsummary-related functions}
